Refactor DashboardStatisticsResponse to use double for TimeSpentHours and update StatsService to round totalTimeHours. Add formatHoursToTime function in DateService for better time formatting in Dashboard.

+++ backend/noava/noava/DTOs/Statistics/DashboardStatisticsResponse.cs
@@ -4,7 +4,7 @@
     {
         public int CardsReviewed { get; set; }
         public double AccuracyRate { get; set; }
-        public int TimeSpentHours { get; set; }
+        public double TimeSpentHours { get; set; }
         public DateTime? LastRevieweDate { get; set; }
     }
 }
